backend/requirements.txt
# Core dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
psycopg2-binary>=2.9.10
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# Phase III: AI Chatbot
openai>=1.50.0
httpx>=0.27.0

# File Upload & Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
